fix(email): preserve paragraph spacing in sent emails
The compose editor renders each paragraph with a visible vertical gap
(my-4 in the markdown theme), but flattenConsecutiveParagraphs joined
consecutive paragraphs with a single <br> in the outgoing HTML, so
recipients saw all paragraphs collapsed onto consecutive lines.

Join non-empty paragraphs with two <br>s so a paragraph break renders
as a blank line, matching what the sender sees. Extract the helper into
its own module and add unit tests.
