+++ apps/web/src/features/block-email/util/prepareEmailBody.ts
@@ -24,6 +24,7 @@ import {
   type LexicalEditor,
   type LexicalNode,
 } from 'lexical';
+import { flattenConsecutiveParagraphs } from './flattenConsecutiveParagraphs';
 import type { ReplyType } from './replyType';
 
 export function clearEmailBody(editor: LexicalEditor | undefined) {
@@ -433,64 +434,6 @@ function applyMediaScale(container: Element) {
     if (height > 0)
       el.setAttribute('height', Math.round(height * scale).toString());
   });
-}
-
-function flattenConsecutiveParagraphs(container: Element) {
-  const paragraphs = container.querySelectorAll('p');
-  const groups = [];
-  let currentGroup: Element[] = [];
-
-  for (let i = 0; i < paragraphs.length; i++) {
-    if (i === 0) {
-      currentGroup.push(paragraphs[i]);
-      continue;
-    }
-
-    // Check if this paragraph immediately follows the previous one
-    const prev = paragraphs[i - 1];
-    if (prev.nextElementSibling === paragraphs[i]) {
-      currentGroup.push(paragraphs[i]);
-    } else {
-      // Start a new group
-      groups.push(currentGroup);
-      currentGroup = [paragraphs[i]];
-    }
-  }
-
-  // Don't forget the last group
-  if (currentGroup.length > 0) {
-    groups.push(currentGroup);
-  }
-
-  // Combine each group and replace in the DOM
-  for (let i = 0; i < groups.length; i++) {
-    const group = groups[i];
-    const div = document.createElement('div');
-
-    for (let j = 0; j < group.length; j++) {
-      const p = group[j];
-
-      const isEmpty =
-        !p.textContent?.trim() &&
-        !p.querySelector('img, video, iframe, canvas');
-
-      if (p.childNodes.length) {
-        div.append(...p.childNodes);
-      }
-
-      if (j < group.length - 1 && !isEmpty) {
-        div.appendChild(document.createElement('br'));
-      }
-    }
-
-    // Replace the first paragraph with the combined div
-    group[0]?.parentNode?.replaceChild(div, group[0]);
-
-    // Remove the rest of the paragraphs in this group
-    for (let j = 1; j < group.length; j++) {
-      group[j].remove();
-    }
-  }
 }
 
 export function prepareEmailBody(
